Added offsetHeight option
Added offsetHeight option to accommodate situations where the parent div is set to position:relative and and a child element is position:absolute. For example, a WordPress text widget with a button that needs to be aligned to the bottom of the parent div. offsetHeight should be a positive integer.